Add a version of recurrence which doesn't use mod
This version is slightly more cumbersome to use because
it doesn't rotate the buffer so the vector given in each
iteration is rotated. For that reason each iteration has
access to an int which shows how much the vector is rotated.
The upshot is that there is no need for the costly mod
operations